requirements.txt
st-annotated-text
torch